Add libssl-dev dependency to sentry module
The python module sentry apparently depends on libssl-dev. This
error is only triggered when one adds the sentry role before the
initial vagrant up (I imagine something else installs it by default).

The question marks are there because cffi and cryptography are not
direct dependencies of sentry and pip is supposed to install direct
dependencies only so I am not sure why these packages need to be
installed in the first place.

Change-Id: I7216fd8641605b82f5bb2c332517b328ee857e0a