Allow use of cachefilesd to speed up file system operations
Add a new configuration option that can be used to enable use of the
cachefilesd service. This service runs on the VM and caches files
accessed from the host NFS server locally on the VM's native filesystem.
This can be enabled with `vagrant config nfs_cache yes`. It is left off
by default for now to allow further testing.

References:
* http://chase-seibert.github.io/blog/2014/03/09/vagrant-cachefilesd.html
* http://linux.die.net/man/8/cachefilesd

Co-author: Bryan Davis <bd808@wikimedia.org>
Bug: T78733
Change-Id: Icdb9ca07cfe81ba70ca924ce2459fcd2f277e684