Add config to run jobchron service
Adds an init config, based on the production jobchron.conf, to run the
jobchron service which is required for delayed jobs used in CirrusSearch
and other places.

Additionaly, as suggested in the ticket, changes the jobrunner and jobchron
to use php5 instead of hhvm.  This results in a memory usage of ~15M per
service rather than the 65M-70M used by hhvm.

Bug: T92278
Change-Id: I1dab95c58d002191671e99ee2b9c80aeff22096b