Introduce the service class and service::node define
WMF services written in Node.JS which use service-runner as their
supervisor or are based on the service template have plenty of things in
common - common configuration file structure, run parameters and
installation steps. In order to minimise copy/pasting and increase code
reusability, the service class and service::node define are introduced.
They cover most of the shared installation steps and scripts, while
still allowing individual services to be separately configured.

As a first step and example, the restbase module is converted to use
service::node. Other services, such as graphoid and mathoid, will follow
this principle in subsequent patches.

Note: this patch also adds the HOME environment variable to the
npm::install and npm::update defines, which allows NPM to work correctly
even when invoked via Puppet. Additionally, --no-bin-links is used for
both install and update to prevent crashes when the host's FS doesn't
support them.

Bug: T90315
Bug: T100202
Change-Id: Ia38d872718abff31c6523d20fa7144ba779076a1