Update and restart services during git-update
Thus far, running vagrant git-update would update the git repos of
services living in /vagrant/srv. This patch enhances its behaviour by
also performing a dependency update and a service restart. For a service
to do that, it must declare a service::gitupdate block specifying which
updates need to be carried out and whether the service should be
restarted. A configuration file for each service is then placed in
/etc/mw-vagrant/services/ and later read by run-git-update. This patch also
updates all of the service modules to use it.

Note: this patch also contains a fix for the phabricator module,
instructing it to connect to MySQL via the TCP socket, not the UNIX one
(the latter has been disabled in the new HHVM version packaged by WMF).

Bug: T101749
Change-Id: I7ec1d2eaa3a003d4c3bfb80a29fa11fb64506407