Create 'suppress' group and put Admin in it
This uses the 'suppress' group name, which has built-in
core i18n (though it doesn't exist in core by default).

Added support in mediawiki::user for an ensuring an account
is in an arbitrary list of groups.  Groups can be added
after the account already exists (even on later puppet runs).

However, groups are not removed by the resource.

Bug: T113641
Change-Id: I8976d5db8b21c2d24b6cb94218b7beb0a3b80851