Enable all puppet-lint tests except 80chars
We don't need all of the legacy cruft of the operations/puppet repo. The
only check that we choose to avoid is the hard limit of 80 characters to
a line. Technically we could deal with this using `lint:ignore:80chars`
comments but they look ugly. We can just trust that this won't be abused
too badly (or clean it up occasionally if it is).

Change-Id: Iaa0ec9765766e17ccca45489d3f727c999108dfa