Replace php::composer::require with mediawiki::composer::require
Replace the unused php::composer::require define with a new
mediawiki::composer::require. The new define collaborates with
composer-merge-plugin to provide a collection of composer.json files
that can be easily managed by Puppet.

Known issues:
* ensure => absent has no effect with composer-merge-plugin v1.2.1:
  https://github.com/wikimedia/composer-merge-plugin/issues/51
* prefer was dropped as it does not seem possible to implement it
  this way. I don't see a use case for it anyway.
* custom directory support was dropped. It would be possible but
  complicated to support it (multiple fragment directories,
  multiple update commands...) and I don't think there is any strong
  use case for it.

Can break in ugly ways if the composer update command breaks for some
reason - since it's notify-based, it will not be run again, and the box
will be left with the role enabled and the dependency not installed and
further provisioning runs do not fix this (disabling and reenabling the
role does). I don't see a way around this.

Bug: T107612
Change-Id: I0a788df3f0456d6e4aab8852c7e98d29ed853489