Add provider based hiera settings and use with LXC
Add a new 'provider' level of hiera lookup that allows setting different
hiera settings based on the Vagrant provider used to create the VM. The
first concrete use case for this is disabling cachefilesd inside the VM
when the provider is LXC.

Also allow specifying the 'environment' hiera level via a 'MWV_ENVIRONMENT'
environment variable. This will allow the Puppet configuration for
running MediaWiki-Vagrant in the Wikimedia Labs environment to switch
the environment similarly to the way that the `labs-vagrant` command
does.

Bug: T107146
Change-Id: I36a0c7af1cd4125e4b699933c5d5bb5fb766ba12