Fix 'no implicit conversion of Pathname into String'
Fixes a bug introduced in I2aabec28a97dfe89ad371d4dde0c05351bdba69a
that passes a `Pathname` instead of a `String`. Usually passing a
`Pathname` to such methods is safe as they rely on `to_path` to
normalize input, but this appears to have exposed a bug somewhere in
Vagrant or its packaged Ruby dependency.

Bug: T111980
Change-Id: I70b07011b27f4a6afb3e5c9cd6b11bddde1cf24d