Require mysql before initializing oozie database
When setting up the analytics role it bailed as unable to connect
to mysql. The passwords were set just a few lines after the failure.

The analytics role still doesn't seem to work, it fails later unable
to start the hadoop-yarn-resourcemanager service, but this patch gets
it moving along at least.

Change-Id: I81140fc449e15f6f19c251a6f7440faf9522f22a