Force usage of cdh zookeeper
Both the cdh and dfsg packages for zookeeper are 3.4.5, but d comes
after c so it wins. The pin priority needs to be > 1000 to force the
downgrade.

hadoop-yarn-resourcemanager will not start with the dfsg zookeeper
installed, not sure why. installing the cdh version to match the
other packages gets it working again.

Change-Id: I411e228093d948651eaf5c541fe718ca516c9d8c