Apache Traffic Server role
This role only aims to cache thumbnails generated by thumbor,
it doesn't try to cache articles. The rationale is that we might want
to put ATS in front of the upload cache in production.

Alternates work, with the different sizes of the same thumbnail
stored under the same cache object, and all getting purged if
one of them is.

The feature set is on par with the Varnish role, expect that
ATS only treats thumbnails.

Bug: T112949
Change-Id: I289825e4b838a187dc6baf593c90102da2529df8