VE: Require RESTBase instead of Parsoid and expose the needed ports
This patch switches the visualeditor role to use directly RESTBase
instead of Parsoid. It also makes Vagrant expose the ports of the
services used by VE - Citoid, RESTBase and Parsoid.

Also in this patch: expose Parsoid's port when its role is enabled. Do
the same for RESTBase, since it uses Parsoid behind the scenes.

Bug: T111223
Change-Id: Iffeb34085ca51e29bc2768acdc631f4bbee2f601