RESTBase VRS: No Parsoid compatibility
When the RESTBase Virtual REST Service was introduced in mediawiki/core
all of the VRS clients were using Parsoid's v1 API layout. However,
that's no longer the case - we have switched them all to RESTBase's API.
That means that they should be free to use all of RESTBase's public API.
The configuration option `parsoidCompat` prevents that as it expects
users to provide only valid Parsoid v1 URLs, so set it to false.

Change-Id: I8f8eade3e8d457638e8ebf97d79734b05c9e55ac