Allow HTTPS port to be configurable
With more than one vagrant box using the https role, they become
unusable due to port conflict. This patch allows configuring the
HTTPS port via 'vagrant config https_port <port number>'.

This is a poor way of doing it as it means the https role will
not work without manually configuring the ports if you have
multiple vagrant boxes (even if only one of them uses the https role
and/or the others are not running at the moment). Not sure how that
could be avoided though.

Change-Id: I6f1632d463c87ad35325bd2af346353d6cc06ece