VE: Set wgVisualEditorRestbaseURL correctly
wgVisualEditorRestbaseURL is used by the client-side VE code to issue
requests to RESTBase from the browser. don't hard-code it to point to
localhost, but instead use the actual domain served by RESTBase (and
MediaWiki) so that labs instances can profit from it as well. Note that
for this to work in labs-vagrant your instance needs to expose the port
7231 to the outside world.

Change-Id: I6a2d8389e67f343efb70275b65b50423584e2aa3