service::gitupdate: Remove node_modules/ before installing the new ones
Most of our Node.JS services use the '^x.x.x' format for specifying the
dependencies' versions to use. When `npm install` is run, it doesn't
check for newer versions if the currently-installed module satisfies the
version. Hence, we need to force npm to install the new versions by
removing the old module dependencies first.

This patch also brings several improvements to vagrant's git-update
command. First, it exposes the environment variables needed by NPM
during the installation process (much like during the initial npm
install on role enabling), allowing for binary modules to be built.
Secondly, it clears the NPM cache before the update process; NPM uses
the cache to compare the versions, which is deadly for large modules
containing binaries as that is known to cause NPM OOMs. Finally, do
not proceed to restarting the service if the update of dependencies
has failed.

Also, update the RESTBase configuration to the current master state.

Change-Id: Ie4769e123850d4ff4da24a22834b2c5d3d03586f