Proxy RESTBase through Apache
Add Apache config to proxy requests for /api/rest_v1/ to the local
RESTBase service and VisualEditor settings to use that entrypoint.

Several small issues in the RESTBase and VisualEditor roles are also
cleaned up. Most notably the extensions and settings for thiese roles
are now only provisioned on the primary wiki as the RESTBase
configuration the roles provide does not currently support multiple
wikis.

Bug: T119731
Change-Id: I7f981d477b9ae25314c1c6ce1903f787e62faf49