Do not set $::port_fragment in labs environment
MediaWiki-Vagrant running in the labs environment will be behind an
external proxy that relays ports 80 and 443 to port 8080 on the Labs
instance. The instance will then in turn forward the port 8080 traffic
to the guest VM. MediaWiki needs to be configured to omit a port number
from generated URLs so that external clients interacting via the Labs
proxy will function correctly.

Change-Id: Ic5a89dc9b2b43e8efd5e528c0f4433d1411b587c