Create kafka role and utilize it from CirrusSearch
This isn't the best possible implementation, but it seems to get
the job done to allow me to easily test messages produced in
avro and shipped to kafka.

It would be nice if this wasn't on by default, but i didn't really
want to make another role cirrussearch-kafka or some such. Open
to ideas.

A kafka role might also be useful for integrating things necessary
to test EventBus in mw-vagrant

Change-Id: I2d9da7ee61be8008137b1f2c30b3bcc23198416c