Fix CentralAuth issues where e.g. Admin might not be fully attached
Run migratePass0 as needed to keep track of users
that need to be attached.

Use attachmissing to request that any existing local accounts
be merged even if global already exists.

In unless, check that it's fully migrated, not just that it exists.

TODO: Make sure all the users are actually attached.  I think
this might require setting the email.


Bug: T126480
Depends-On: I81764696d2277c0d11d4440aa484a2c5ca24f354
Change-Id: I30d78e1dfc7923872824852379596959cfbe43fe