Use protocol-relative wgServer and https role changes
wgSecureLogin was being disabled due to not being a protocl-relative
URL.  This can not be fixed within the https role itself because
detectServer is running in CommonSettings.php after all settings.d
are done.

So instead we change how wgServer is determined in CommonSettings.php
(regardless of roles).

Also, if the https role is enabled (we could look at doing it
regardless of role if this works well, though it needs to run late due
to wgAssumeProxiesUseDefaultProtocolPorts), make the wgServer values
in wgConf port-relative.  Basically, this uses the canonical host name
and a protocol-relative host, but the port of the current request.

Bug: T126378
Change-Id: I4a2d7d56c8cafbfde4751cbb176ea0c280035ac5