Remove HHVM fss plugin
It's missing a dependency on the latest HVVM. It's
no longer used in production.

Had to remove the "current" symlink as well, since
hhvm --version no longer returns the API version
number we need in order to know which directory
is the current one. The value is hardcoded the same
way in production puppet for the same reason.

Bug: T129343
Change-Id: I97114a7da98d5f4b5afb7c1dd3901489374b8734