HHVM: Fix ordering so that hhvm-fss can be removed
Follow up to I97114a7 which fixed the root problem for provisioning
a new VM.

Clean up some legacy cruft in ::hhvm and add new and improved legacy
cruft to deal with the upgrade to hhvm 3.12.1+dfsg-1. The new version of
hhvm is not compatible with hhvm-fss and new versions of that package
aren't being built since it is no longer used in WMF production. There
is a chicken and egg problem with upgrading to the new hhvm package on
an existing instance.

* Purge hhvm-fss if installed *before* updating hhvm
* Update config files *before* installing/updating hhvm so that
  references to hhvm-fss are removed. This inverts the logic applied to
  fix T115450.

Other cruft cleaned up:

* Remove env::alternative dependency on hhvm startup as symlink
  management was removed in I97114a7.
* Remove force=>true hack for T87478.
* Remove unused hdf config template and old comment about it.
* Remove legacy /etc/hhvm config cleanup. The HHVM package still
  provisions /etc/hhvm/server.ini, but we can live with that.

Bug: T129343
Change-Id: I999b8dc7d402d5e2b986be4659c6d93825548613