multiversion: fix wgConf global extraction
Iaf9bf8d tried to improve error messages when mwscript was either not
used or had an invalid wiki supplied. In the process however
extractAllGlobals() was moved to a point before $wgDBname was properly
set. Additionally, the protections that the prior patch attempted to
provide do not actually work as intended. The $wgLocalDatabases count
guard contition will never be true due to improper placement relative to
the MWMultiVersion.php require. The MWScript.php guard condition will
not be reached due to the actions which happen in MWScript itself.

Bug: T123143
Change-Id: I26eb68e9961b1c1c30728565938a097c67b48139