Allow global extension to be disabled per-wiki
Some extensions (e.g. AuthPlugins) are mutually exclusive. Roles like
role::centralauth that setup an extension globally for the
MediaWiki-Vagrant managed wiki farm are the norm due to simplicity.
Another role however may want to setup a single wiki that deviates from
the norm for the rest of the wiki farm.

This change introduces a method for disabling an extension that has been
installed globally on a per-wiki basis using the new $wmvExtensions
global PHP variable. This variable is expected to be an array with
extension names as keys and booleans as values. If an extension is not
listed in the array or maps to true then it will be loaded. If it maps
to anything else it will not be loaded.

::mediawiki::wiki is extended to allow providing a hash of additional
$wgConf settings for the wiki. This can be used to set $wmvExtensions
flags or change any other setting currently handled by $wgConf.

Bug: T128501
Change-Id: I2ad8cc316d99c27e1d92c2bb776a4224e57ec08c