Introduce foreachwikiwithextension
Follow up to I2ad8cc3. Now that it is possible for a wiki to opt-out of
enabling a globally installed extension we need a way to also exclude
those wikis from `foreachwiki` scripts. In the Wikimedia production
environment, this would be done using `foreachwikiindblist` and manually
curated dblist files. In MediaWiki-Vagrant we'd like thing to be a bit
more automated. This change introduces a new `wikihasextension` helper
script that can generate a dblist of wikis with a given extension
enabled. This is paired with the new `foreachwikiwithextension` script
which is a `foreachwiki` work-alike that takes an extension,
a maintenance script and optional additional arguments.

The extensions active for a wiki are tracked in a new
`$wmvActiveExtensions` global variable as part of the actions of the
mediawiki::extension Puppet module. Existing uses of `foreachwiki` that
depend on a particular extension being installed have been updated to
use `foreachwikiwithextension` instead.

All of this will fix problems found when combining role::centralauth and
role::ldapauth in a single MediaWiki-Vagrant VM.

Bug: T128501
Change-Id: I7b2bf3d25ec148a0e949159178bf2616b5739c82