Hook Thumbor into the custom rewrite.py Swift middleware
This is almost the same rewrite.py used in production. Requests now
go from Varnish to Swift. Which when it doesn't have the file it needs,
requests it from Thumbor, which generates it on the fly.

Thumbor is responsible for storing the thumbnail in swift and serving
it to rewrite.py for it to bubble back up the request chain.

I had to disable some strange hardcoded proxying (where the image
scaler proxies itself?) and add support for headers that are meaningful.

The xkey header in particular is set up so that it's generated in thumbor
and stored in swift permanently. This way when an existing thumbnail
has to be pulled from swift, it will have an xkey. As a result, varnish
never needs to compute xkeys, it only consumes them. Which helps
mitigate encoding issues that might happen.

Overall this simplifies VCL greatly, which is kept dumb. Note that this
doesn't handle Swift shards yet, not multiwiki setups. Both of which
will be required for something like this to work in production.

Bug: T126521
Change-Id: I13d53401b1c4875ff74a111336fb58f35570d862