Fix hhvmsh by porting mwrepl
This ports over a modified version of mwrepl from operations/puppet.

This fixes hhvmsh (it was broken for multi-wiki setups, due to
b48d739bf6e444cf17db754c00f01abc46f6f7de , since the startup document
was /vagrant/mediawiki/index.php (not multi-wiki aware).

This instead uses the mwrepl style, which does not use a startup
document.

This also allows accessing a shell for any database, not just 'wiki'.

hhvmsh is a symbolic link to mwrepl for back-compat.

hhvmsh did work a little differently (it used debug-host).  If anyone
is actually relying on that, we could put it back.  However, we would
then have three implementations, mwrepl, phpsh, and hhvmsh.

Bug: T127759
Change-Id: Idffcbbb9501d41113e39bca9a123e93c222ab270