git-update: Update the deps only when required
Up until now, `vagrant git-update` would blindly try to update all of
the services' dependencies regardless of whether that's actually needed
or not. This patch makes the git-update script inspect the list of files
that have actually changed, and if npm, composer or pip config files
have, it updates the dependencies. Otherwise, no such action is taken.

Bug: T125765
Change-Id: I65bdc56ec7415237652e394a4634637b6e7c2b25