Install logstash contrib plugins by default
Logstash currently fails to start when the elk role is enabled due to
utilizing logstash-filter-prune which is part of contribs and not in the
standard distribution. In newer version of logstash it would make sense
to create a logstash::plugin resource, but in the current 1.4.x version
contrib is the only possible plugin, so let's just install it by
default.

Change-Id: I94c111bcea73ac2f5fc0e71ba8f3bebe59614313