service::update: Tell pip to upgrade the packages if possible
I was running into the problem of the eventbus proxy service not being
up and running after a vagrant git-update. It turns out that it was
failing because the requirements weren't updated, while the code was
counting on dependencies being there. To avoid that, pip needs also to
try to upgrade the dependencies when possible.

Change-Id: I63aabdf893f57656fb7c74762c25063cb748c5ac