Make Swift sharded
This makes the Vagrant Swift instance use sharded
containers.

It also lets Swift do the url rewriting via rewrite.py
instead of reproducing the same logic in Apache's
proxying.

It makes a slight modification to the stock rewrite.py
in order to allow for single-character shards. Otherwise
it would force us to use 256 shard per logical container
which is really overkill for the VM and time-consuming
to provision. MediaWiki already supports single-character
Swift sharding out of the box.

The container creation check is also simplified by relying
on a file that already exists, instead of creating a temp
file every time, which had a tendency to disappear and
slow down provisioning occasionally.

A mechanism to restart swift on provisioning after
config changes is also introduced.

Bug: T126496
Change-Id: I299e2e0d38ccd426a647af6346a343a8dcf268f6