Do not recognize unkown DB names just because they end in 'wiki'
MediaWiki tries to parse DB names based on the assumption that they
are composed of a language code plus a common suffix (*wiki,
*wiktionary etc) and there are too many language codes to easily
enumerate. Neither assumptions is true for vagrant so we can just
define a full list of accepted site names. That avoids problems
the would arise when a wiki is disabled and its name ends with
another wiki's name (so the disabled wiki would be considered to
exist and that would break WikiMap amongst other things).

Change-Id: I18701d9e812cb27b997664ed04a683cbda7cd456