jobrunner: fix for vms with multiple default vhosts
Add a 'Host: vagrant' header to jobrunner's curl command to help Apache
select the MediaWiki vhost. This will fix a bug for VMs that have
enabled roles which add new vhosts that bind to _default_:80 and will
also work as expected on "normal" VMs that are only using the MediaWiki
vhosts.

Change-Id: I7103b28cda87fb9b726744971cab95bcbfec9cd1