Allow hiera config of jobrunner groups
Interoduce a mediawiki::jobrunner::groups hiera variable that is
expected to contain a mapping (YAML for dict/hash/object) describing the
jobrunner groups to configure in /etc/jobrunner.json. Commmon.yaml
recreates the current default config, but a user can override this in
their local.yaml to change the number of runners or introduce separate
processing queues for certain jobs types as is done in WMF production.

Inspired by https://www.mediawiki.org/wiki/MediaWiki-Vagrant#Job_queue
instructions about hacking the config file in the VM which is gross.

Change-Id: Ic5d1e0e9aae49c7df5ef220f7b600fd93267fc23