service::node: Parse the configs when merging and update config
The merge_config parser function used to just take the two config
strings and concatenate them. However, that creates problems when there
are duplicate keys (when the service wants to override a setting).
Therefore, the configs need to be parsed and merged as hashes.

As of v0.3.2, services based on service-template-node need extra config
settings that point to the MW and REST API endpoints. This commit adds
them.

Also, clean up the config for the mobilecontentservice module.

Change-Id: Ifad5b3f772e6402e3b2eec533926324410cae53a