NPM: pin nodejs to always be installed from the nodesource repository
The Wikimedia APT repo is pinned to > 1000, which means that APT will
try to downgrade a package when its version is newer than the installed
one. We don't want to do that for nodejs, so pin it to the nodesource
repository, which will always give us the newest node 4.x.

Bug: T148348
Change-Id: Ib2208d47b48c06b7a846e763390eac9d09674d2a