Drop the vbguest plugin and sideload the mediawiki-vagrant plugin
This started as a patch to remove the vbguest plugin due to the various
problems that it has caused with apt-get after an initial Puppet run.
The vbguest plugin hooks into the VM startup sequence before local
shares are mounted and tries to run apt. This is fine until the first
Puppet run has happened in the VM. Our Puppet code messes around with
the apt cache directroy structure to assist with caching across VM
instance create/destroy cycles. These changes leave the apt
configuration pointing at a /vagrant/cache/apt directory that doesn't
exist except when the local shares are mounted in the VM.

This causes enough problems for people that I have a paste
(<https://phabricator.wikimedia.org/P2955>) to point them to as a quick
fix.

The scope grew however when Vagrant 1.9.0 was released with internal
changes that break our local plugin install as well. Our plugin was
loaded directly via the Vagrantfile before an internal change was
introduced in Vagrant 1.6.0 that made that impossible. A new way to work
around that regression was introduced in Vagrant 1.7.0 and that method
of loading a plugin from local disk continues to work in Vagrant 1.9.0.

The end result is that setup.{bat,sh} no longer installs the vbguest
plugin and our local mediawiki-vagrant plugin is read directly from the
local git clone rather than being compiled and installed.

The one desirable feature that has been lost in the current
implementation is forcing the user to run setup.{bat,sh} on initial
install and periodically there after when the major version number of
the plugin is bumped. I'm sure we can find a way to add this back by
storing some metadata in the .settings.yaml file.

Bug: T151928
Change-Id: I07bc0e1fbf73d665e439dcdbf1805acf6ecc2746