Clone only the master branch of mediawiki/core.git by default
We've been seeing some issues with bad commit references causing the
initial git clone of MediaWiki to fail. Various debugging attempts point
to commits on deployment branches that may have somehow been lost in the
origin repo. Normal MediaWiki development happens on the master branch
and not a release branch, so we can work around this issue by limiting
the initial clone to a single branch.

Bug: T152801
Change-Id: I4939712110c0ac9286b668ca7a4f86440fb28e4f