Use puppet-syntax gem for syntax checking
The puppet-syntax gem lets one validate puppet manifests, erb temlates
and hiera yaml file.  That lets one easily validate locally before
sending a patch for review and would let us remove the Wikimedia Jenkins
job erblint-HEAD and pplint-HEAD.

Add the gem puppetlabs_spec_helper which provides rake tasks wrapping
around puppet-syntax. Flag it with require: false, there is no need to
have bundler to always require it, it is sufficient to just require it
in the Rakefile.

The puppetlabs_spec_helper provides:

    rake syntax             # Syntax check Puppet manifests and templates
    rake syntax:hiera       # Syntax check Hiera config files
    rake syntax:manifests   # Syntax check Puppet manifests
    rake syntax:templates   # Syntax check Puppet templates

Make the 'syntax' task a prerequisites of the 'test' task in order to
have CI to run it for us.

Note: the submodules are validated, though they each have their own
validation as well and overall it is probably a good idea to make sure
they have proper puppet manifests.

Bug: T154894
Change-Id: I29de3b2b49db245a46b1acdc862701f6b872a54d