Add role for ORES and Wikilabels services
Updates the virtualenv resource to allow editable installs,
and adds ores_services role which installs ORES and Wikilabels
as editable pip packages, and sets up upstart services and vhosts
for them.

Making the ores role make use of the ores_services role when both
are enabled is left to a later patch.

Bug: T159105
Change-Id: Iea462d8316bb5284a382bf4f6f04c44c221362cc