Use a bash script to handle elasticsearch plugin install/removal
Versioning of elasticsearch plugins is a complete pain. Whenever there
is an upgrade, either minor or major, it ends up causing problems with
vagrant provisioning all the correct pieces.

The difficulty with the previous solution, a Facter that provides the
elasticsearch version, is that it detects the version *before* the
puppet catalog has been applied. This means a resource that puts the
wikimedia repo into place is not taken into account, and the wrong
version of elasticsearch + plugins can be installed.

This tries to move that complication into a custom bash script, rather
than managing it from puppet. That allows the elasticsearch plugin to
be set to the 'latest' version, and the plugin script to handle making
sure the installed version of elasticsearch matches the installed
version of the plugins.

Change-Id: Ieef00679281884a411569828867a2e1b5939a1c8