Make puppet order deterministic across hosts
From time to time we hit a hard to debug issue where Puppet's default
title-hash ordering produces a different order on someone's VM
that then leads to a dependency cycle or resources otherwise being
applied in an order that is unexpected. Puppet 3.3.0 introduced an
optional `manifest` ordering where the default order of resources is
a top to bottom reading of each manifest as it is encountered. Forcing
this ordering should help reduce bugs by improving the consistency of
ordering from host to host and vm to vm.

Bug: T161127
Change-Id: I1981d40242b03d6e28bdf2a91172045a5a6ea5d3