mwv: separate require_package usage from File precondition
Since 655a94a require_package actually creates an ordering dependency
between the package and the enclosing class. This can cause problems if
the enclosing class also in some way creates a 'before' ordering
dependency on Exec[apt-get update] or anything else that is
a precondition of installing the package. The particular cycle being
broken here is one that is only triggered when NFS shares are used and
::mwv places a 'before' condition on all File resources.

Bug: T161127
Change-Id: I6af63f8d6f01db14cc784fe735877c6ce5652a92