Add elastic quirks to vagrant
I think vagrant should set this value as it prevents default install from
running properly. This option being rather obscure we should set it to
a convenient default.
When elatic v5.3 or later is uploaded to the repo we can remove it as
it will generate warnings in elastic logs.

Change-Id: I8202b1dcb1271fc8163d17641e54f9a7b1aea697