Puppet: Downgrade npm to version 2
After we've upgraded node.js in Vagrant to version 6,
it brought us npm version 3. However, version 3 has a
weird bug that doesn't allow properly installing dependencies
with --no-bin-links.

I couldn't find a workaround for the bug or any reference that
it's been addressed in version 3, so let's stick with version 2
for now.

Also, we don't need nodejs-legacy ane more or older legacy npm
to be installed, so deleted puppet that used to install them.

Bug: T162468
Change-Id: I0c9c2573c0b2163976f8541fa86637f8a99149d4