Switch to git from jessie-backports
We have been using an Ubuntu PPA to get less ancient git binaries since
we were on the Trusty base image. This was most recently installing
a 2.13.0 version of git. The hacks to use a PPA on Jessie are fragile
and the more robust jessie-backports apt repo has version 2.11.0 which
is new enough.

This change will not alter the installed git version on any existing VM.

Change-Id: I476928e95bd808072208144d1284995bda6906b2