Simplify setup.{bat,sh}
The complexity of finding the Vagrant ruby interpreter and running
custom ruby code is unnecessary since we switched from full plugin
install to sideloading in c8ca590 and stopped installing external
plugins. Instead we can just call `vagrant config --required`.

This change was prompted by a report on irc that the
"..\..\embedded\bin\ruby.exe" path is invalid with the latest version of
Vagrant for windows.

Bug: T171380
Change-Id: I161be221ff1c51eae029c0897d20de7a3271b7ad