rspec-puppet for all the roles
Bump rspec-puppet 2.5.0...2.6.9 to catch up with a lot of fixed issues
and new features.
Update rspec but stick to 3.5.x for now (ran bundle update rspec-puppet)

Under /puppet add the boilerplate to support rspec-puppet. The usage is
pretty straightforward:

  bundle install
  cd puppet
  bundle exec rake spec

That will look up for all manifests in /puppet/modules/role/manifests/
and attempt to compile the catalog for each of them. That is similar to
what we do on operations/puppet.

rspec-puppet always set Puppet[:app_management], there is a bug in
Puppet 4.8.x which then causes defines to trigger a syntax error when
they have a 'site' or 'consumes' parameter such as our
apache::site_conf.
To workaround that, mock Puppet.settings so that [:app_management]
always return false.

https://github.com/puppetlabs/puppet/commit/52ff5258a9ce5bc3038f20c913823d370448a7c0
https://tickets.puppetlabs.com/browse/PUP-6917

Similary, puppet 4.8.2 does not set 'systemd' as the default Service
provider since it does not recognize stretch. Workaround that by setting
the default provider in a :pre_condition.
The puppet Debian package is not affected since it comes with a patch:
https://anonscm.debian.org/cgit/pkg-puppet/puppet.git/tree/debian/patches/0005-use-systemd-as-the-default-service-provider.patch?h=jessie-backports

Under the hood, it uses rspec, and one can pass it options via the
SPEC_OPTS environment variable:

 cd puppet
 SPEC_OPTS='--help' bundle exec rake spec

Set spec statuses persistence so one can easily rerun solely the failed
tests via:

 cd puppet
 SPEC_OPTS='--only-failures'

The processorcount is always set, that is used by a few roles:
- role::analytics
- role::hadoop
- role::hive
- role::hue
- role::oozie

Pending:
- should set facts using rspec-puppet-facts

Bug: T183570
X-Notable-Event: Merry-Christmas
Change-Id: I5e286b88d1dbecb10860afbf2ac99704b33c5e58