thumbor: `nproc` => $::processorcount
When compiling role::thumbor with rspec on a host that miss 'nproc'
(typically with rspec-puppet on a Mac), the catalog compilation fails
with:

Error while evaluating a Function Call, Failed to parse inline template:
No such file or directory - nproc
  at puppet/modules/thumbor/manifests/init.pp:124:35

Use the processorcount fact instead.

Then $ports is an array of Integer, however a define expect a String as
name:
  Evaluation Error: Illegal title type at index 0.
  Expected String, got Integer
  at puppet/modules/thumbor/manifests/init.pp:126:24

Convert the sequence_array from an array of Integer to an array of
String.

Bug: T183462
Bug: T183570
Change-Id: Iacdf2c95f4286374bcafc98a4c0cff5c970e529c