mysql: Provide default charset for MediaWiki tables
Provide a default charset of 'binary' for auxiliary tables provisioned
for:
* centralauth
* contenttranslation
* externalstore
* globalblocking

The global default charset for the Stretch MariaDB install is utf8mb4
which causes issues for indexing when ROW_FORMAT=DYNAMIC is not provided
in the CREATE TABLE declaration. Globally setting ROW_FORMAT=DYNAMIC is
not possible in MariaDB <10.2.2.

Bug: T183679
Change-Id: Icbf3d397af465c1b09a43963dbe28af0f269b53f