rake: Make `rake test` reentrant
`rake test` leaves a bunch of files behind in tmp/testenv. If these
files are not removed before the next run of `rake test` the linter will
fail with a long list of duplication messages. Adding a :clean step to
the start of the test run will take care of this by ensuring that the
prior test run's files are removed. This should have no effect on CI
tests, but will make local testing easier.

Change-Id: Iaeb2df655f92943b91f8e2410ae82034d46c53de