Compile the base puppet catalog on CI
Add rspec-puppet tasks at the root of the repository so one can easily
compile the catalogues.
First for a dummy host that has no role applied, and if successful for
each roles. That is done serially which is quite slow.

Add the fast :compile_host task to the default commands being run by CI.
It is fast enough.

Bug: T183570
X-Notable-Event: Happy-New-Gregorian-Year
Depends-On: Iedfd61c022f39d32f6d803a39579bb5d121537ed
Change-Id: I4ca7a2b80d7b0ffb8b3fcef40c498b8c5dfc1ba0