wikimetrics: remove role
Remove the wikimetrics role and related Puppet module. This module was
removed from Wikimedia's operations/puppet.git repo in its f79b155
revision. The code present here has not been updated for systemd, which
means that it does not work on the current Debian Jessie master branch
of MediaWiki-Vagrant or the Debian Stretch testing branch.

Bug: T184154
Change-Id: I5d315c0f0b6be94b5559f0d1afe21704104b1ce1