add supporting shell scripts for dump role and restructure dir locations
Having the shell script that sets directories will make it easier for people
testing 'misc' dumps that ordinarily run out of cron, to the extent that the
databases and/or wikis can be set up in vagrant.

These same bash scripts also always invoke MWScript.php in a hardcoded
directory path, so we symlink the directory containing that script to
the location where the scripts expect it to to be.

The directory layout is now more similar to production, making maintenance
of the vagrant role easier, and making it easier for dumps testers to
get around.

A separate user for dumps is introduced here, just as in production,
so that any issue with permissions shows up in testing.

Also: clean up extra dir permissions we didn't need, add the php-bz2
extension, which somehow was missing, add newline to all.dblists entries

Bug: T201478
Change-Id: I712332138a8549fe9541763b4b8c4087629947b2