Add interface-admin exemption in password policies
In the linked bug people have been complaining about the new password
policies. I have tested this fix on my own copy of the code and it
appears that the interface-admin group has stricter password policies
not allowing the use of the default password where the group exists.
This patch serves to remedy that by overriding the strict policy with
one that allows the default password to be used.

Bug: T203486
Change-Id: I290f0285aaed080269754c58727390cb58b5af01